feat: create unified MatchResultHeader component
- Create MatchResultHeader component with variant prop (latestResult | matchDetail)
- Uses dark gradient background with blur elements from LatestResultBlock
- Displays all match data: score, scorers, competition, season, venue, attendance, referee
- "Latest Result" title and CTA buttons only shown for latestResult variant
- Refactor LatestResultBlock to use the new unified component
- Update match page to use MatchResultHeader for post-match status

Closes #64
